+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## [1.12.2 - 3.2.4]
 
+* fix: correct redstone update 
 * ref: update buildpipeline
 
 ## [1.12.2 - 3.2.3]